requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
openai==0.28
feedparser==6.0.10
newspaper3k==0.2.8
schedule==1.2.0
requests==2.31.0
supabase
python-dateutil==2.8.2
asyncio==3.4.3
httpx>=0.23.0,<1.0.0
